Support vertical reordering
