+++ src/Files.App.Controls/ReorderablePanelContainer/ReorderablePanelContainer.xaml
@@ -3,9 +3,9 @@
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:local="using:Files.App.Controls">
 
-	<Style BasedOn="{StaticResource DefaultSnapPanelContainerStyle}" TargetType="local:ReorderableStackPanelContainer" />
+	<Style BasedOn="{StaticResource DefaultSnapPanelContainerStyle}" TargetType="local:ReorderablePanelContainer" />
 
-	<Style x:Key="DefaultSnapPanelContainerStyle" TargetType="local:ReorderableStackPanelContainer">
+	<Style x:Key="DefaultSnapPanelContainerStyle" TargetType="local:ReorderablePanelContainer">
 		<Setter Property="ItemsPanel">
 			<Setter.Value>
 				<ItemsPanelTemplate>
