+++ tests/Files.App.UITests/MainWindow.xaml.cs
@@ -54,7 +54,7 @@ namespace Files.App.UITests
 					nameof(BreadcrumbBarPage) => typeof(BreadcrumbBarPage),
 					nameof(ResizeVisualPage) => typeof(ResizeVisualPage),
 					nameof(TableViewPage) => typeof(TableViewPage),
-					nameof(SnapPanelContainerPage) => typeof(SnapPanelContainerPage),
+					nameof(ReorderablePanelContainerPage) => typeof(ReorderablePanelContainerPage),
 					nameof(DragSelectionContainerPage) => typeof(DragSelectionContainerPage),
 					_ => throw new InvalidOperationException("There's no applicable page associated with the given key."),
 				});
